requirements.txt
opencv-python
aiortc
av
numpy
python-socketio[client]
onnxruntime
